reorganizando exemplos, atualizando env e editando readme
