+++ environment.yml
@@ -1,5 +1,6 @@
 name: flask_tutorial
 channels:
+  - bioconda
   - defaults
 dependencies:
   - _libgcc_mutex=0.1=main
@@ -29,8 +30,14 @@ dependencies:
     - cycler==0.10.0
     - decorator==4.4.1
     - defusedxml==0.6.0
+    - dnspython==2.2.1
+    - email-validator==1.1.3
     - entrypoints==0.3
     - flask==1.1.1
+    - flask-sqlalchemy==2.5.1
+    - flask-wtf==1.0.0
+    - greenlet==1.1.2
+    - idna==3.3
     - ipykernel==5.1.4
     - ipython==7.12.0
     - ipython-genutils==0.2.0
@@ -72,6 +79,7 @@ dependencies:
     - seaborn==0.10.0
     - send2trash==1.5.0
     - six==1.14.0
+    - sqlalchemy==1.4.32
     - stats==0.1.2a0
     - terminado==0.8.3
     - testpath==0.4.4
@@ -81,4 +89,5 @@ dependencies:
     - webencodings==0.5.1
     - werkzeug==0.16.1
     - widgetsnbextension==3.5.1
+    - wtforms==3.0.1
